+++ src/App.css
@@ -8,6 +8,12 @@ p {
   box-sizing: border-box;
 }
 
+button {
+  cursor: pointer;
+}
+
+/*HOME PAGE*/
+
 /*Header*/
 .select__wrap {
   flex-basis: 19.5%;
@@ -71,7 +77,7 @@ p {
   }
 }
 
-@media (max-width: 768px) {
+@media (max-width: 678px) {
   .specPrices__cards {
     flex-direction: column;
   }
@@ -105,17 +111,13 @@ p {
 }
 
 /*Footer*/
-
-/* .footer_social {
-  margin-top: 20px;
-} */
 @media (max-width: 450px) {
   .footer_btnWrap {
     flex-direction: column;
   }
 }
 
-/*Search-Page*/
+/*SEARCH-PAGE*/
 .search__header_wrapper {
   display: flex;
   padding-top: 80px;
@@ -162,8 +164,7 @@ button.search__header_btn:hover {
   }
 
   .search__header_wrapper .select__wrap {
-    max-width: 24.5%;
-    margin-right: 2px;
+    max-width: 24.6%;
   }
 
   .search__header_wrapper .select__wrap > div {
@@ -188,7 +189,6 @@ button.search__header_btn:hover {
 .DayPickerInput input {
   display: block;
   width: 100%;
-  /* margin-right: 2px; */
   padding: 18px 0 18px 10px;
   flex-basis: 17.5%;
   box-sizing: border-box;
@@ -256,88 +256,56 @@ button.search__header_btn:hover {
   background: #eee !important;
 }
 
+/*Header DAte Inputs*/
 .InputFromTo {
   display: flex;
+  flex-basis: 36%;
   position: relative;
 }
 
-.InputFromTo-to {
+.InputFromTo > div {
+  flex-basis: 50%;
+  margin-right: 2px;
+  border-radius: 0;
+}
+
+.DayPickerInput {
   position: relative;
+  width: 100%;
 }
 
 @media (max-width: 992px) {
   .InputFromTo {
-    flex-basis: 49.7%;
+    flex-basis: 49.8%;
   }
 
-  .InputFromTo .DayPickerInput > input {
-    border-radius: 0 0 0 4px;
-  }
-
-  .InputFromTo-to .DayPickerInput > input {
-    border-radius: 0;
-  }
-}
-
-@media (max-width: 992px) {
-  .InputFromTo {
-    flex-basis: 100%;
-    margin-bottom: 2px;
-  }
-
-  .InputFromTo > div {
-    flex-basis: 50%;
-    margin-right: 2px;
-    border-radius: 0;
-  }
-
-  .InputFromTo .DayPickerInput > input {
-    border-radius: 0;
-  }
-}
-
-/* @media (max-width: 1200px) {
-  .DateInput-wrapper,
-  .DayPickerInput,
-  .DayPickerInput input {
-    flex-basis: 16.4%;
-  }
-}
-
-@media (max-width: 992px) {
-  .DateInput-wrapper,
-  .DayPickerInput,
-  .DayPickerInput input {
-    flex-basis: 24.7%;
-  }
-  .DayPickerInput input[placeholder="Туда"] {
+  .InputFromTo > div:not(.InputFromTo-to) input {
     border-radius: 0 0 0 4px;
   }
 }
 
 @media (max-width: 576px) {
-  .DateInput-wrapper,
-  .DayPickerInput input {
-    flex-basis: 49.3%;
-    margin-right: 2px;
-    margin-bottom: 1px;
+  .InputFromTo {
+    flex-basis: 100%;
+    margin-bottom: 2px;
   }
 
-  .DayPickerInput {
-    width: 100%;
-  }
-
-  .DayPickerInput:first-of-type,
-  .DayPickerInput:first-of-type input {
+  .InputFromTo > div:not(.InputFromTo-to) input {
     border-radius: 0;
   }
-} */
+}
 
 /*Downshift Input*/
 .downshift-input {
   position: relative;
   flex-basis: 22%;
   margin-right: 2px;
+}
+
+@media (max-width: 1200px) {
+  .downshift-input {
+    flex-basis: 20%;
+  }
 }
 
 @media (max-width: 992px) {
@@ -347,5 +315,16 @@ button.search__header_btn:hover {
 
   .downshift-input input {
     border-radius: 0 4px 0 0;
+  }
+}
+
+@media (max-width: 576px) {
+  .downshift-input {
+    flex-basis: 100%;
+    margin-bottom: 2px;
+  }
+
+  .downshift-input input {
+    border-radius: 0;
   }
 }
