+++ src/Header/Input.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 import styled from "styled-components";
-import DayPickerInput from "react-day-picker/DayPickerInput";
 import "react-day-picker/lib/style.css";
 
 import Icon from "./Icons";
@@ -79,25 +78,11 @@ const InputTag = styled.input`
       border-radius: 0 4px 0 0;
       margin-bottom: 2px;
     }
-
-    &#dateFrom {
-      border-radius: 0 0 0 4px;
-    }
   }
 
   @media (max-width: 576px) {
     &#departure {
       border-radius: 4px 4px 0 0;
-      margin-bottom: 2px;
-    }
-
-    &#arrival,
-    &#dateFrom {
-      border-radius: 0;
-    }
-
-    &#dateFrom,
-    &#dateTo {
       margin-bottom: 2px;
     }
   }
