+++ src/Home.js
@@ -1,8 +1,5 @@
 import React, { Component } from "react";
 import { Helmet } from "react-helmet";
-// import { BrowserRouter, Route, Link } from "react-router-dom";
-// import { Grid, Row, Col } from "react-flexbox-grid";
-// import styled from "styled-components";
 import MediaQuery from "react-responsive";
 import Header from "./Header";
 import TopDestinations from "./TopDestinations";
