+++ src/SearchPage/FilterBlank.js
@@ -34,68 +34,11 @@ const Title = styled.p`
   }
 `;
 
-// const Label = styled.label`
-//   display: flex;
-//   width: 100%;
-//   padding: 12px 0;
-//   font-size: 12px;
-//   color: #4a4a4a;
-
-//   & span {
-//     display: block;
-//     margin-left: auto;
-//     color: #a0b0b9;
-//   }
-
-//   &::before {
-//     content: "";
-//     width: 18px;
-//     height: 18px;
-//     transform: translateY(-3px);
-//     margin-right: 6px;
-//     border: 1px solid #a0b0b9;
-//     border-radius: 4px;
-//   }
-// `;
-
-// const Input = styled.input`
-//   &:checked + label::before {
-//     border: none;
-//     background: url(${checked}) no-repeat;
-//   }
-// `;
-
-// const BlockFilters = styled.div``;
-
 class FilterBlank extends React.Component {
   render() {
     return (
       <BlockWrapper>
         <Title count={this.props.count}>{this.props.title}</Title>
-        {/* <BlockFilters>
-          <Input type="checkbox" hidden checked />
-          <Label>Все</Label>
-          <Input type="checkbox" hidden />
-          <Label>
-            Без пересадок
-            <span>5 123 ₽</span>
-          </Label>
-          <Input type="checkbox" hidden />
-          <Label>
-            1 пересадка
-            <span>4 777 ₽</span>
-          </Label>
-          <Input type="checkbox" hidden checked />
-          <Label>
-            2 пересадки
-            <span>4 123 ₽</span>
-          </Label>
-          <Input type="checkbox" hidden />
-          <Label>
-            3 пересадки
-            <span>3 789 ₽</span>
-          </Label>
-        </BlockFilters> */}
       </BlockWrapper>
     );
   }
