+++ src/index.js
@@ -4,7 +4,6 @@ import { BrowserRouter, Route, Link } from "react-router-dom";
 import normalize from "normalize.css";
 import "./index.css";
 import App from "./App";
-import Search from "./SearchPage";
 import registerServiceWorker from "./registerServiceWorker";
 
 ReactDOM.render(
